requirements.txt
Flask>=3.0.0,<4
Flask-Minify>=0.50,<1
gunicorn>=21.2.0,<22
python-dotenv>=1.0.0,<2
psycopg2-binary>=2.9.9,<3
